fix: 删除接口

+++ server/api/table/column.delete.ts
@@ -1,5 +1,5 @@
 import { ObjectId } from "mongodb";
-import { columns_collection } from "./index.post";
+import { columns_collection, rows_collection } from "./index.post";
 import { z } from "zod";
 
 export const id_schema = z.object({
@@ -8,7 +8,15 @@ export const id_schema = z.object({
 
 export default defineEventHandler(async (event) => {
   const { _id } = await getValidatedQuery(event, id_schema.parse);
-  return columns_collection.deleteOne({
+  const column = await columns_collection.findOne({
     _id: new ObjectId(_id),
+  });
+  if (!column) throw createError({ status: 404 });
+  await rows_collection.updateMany(
+    { table_id: column._id },
+    { $unset: { [_id]: true } },
+  );
+  return columns_collection.deleteOne({
+    _id: column._id,
   });
 });
